src/ai/transaction-monitoring-agent/requirements.txt
requests>=2.0.0
google-generativeai>=0.3.0